Created cache for rendered body content
In order to speed up template rendering, the rendered body of a message
will now be stored in body_cache and retrieved if populated.

When we add write functionality we'll have to make sure that any body is
rendered on save to take the onus from the template rendering.  This is
just a stop-gap.